chore(basemodal): align with carbons composed modal
