+++ README.MD
@@ -51,7 +51,7 @@ Note that previously, v1.x required installation of additional packages `carbon-
 You can then import any component that you need by doing the following in your project:
 
 ```js
-import { BaseModal } from 'carbon-addons-iot-react';
+import { ComposedModal } from 'carbon-addons-iot-react';
 ```
 
 ## 📚 Docs
