--- src/components/BaseModal/index.js
@@ -1,1 +1,0 @@
-export default from './BaseModal';
