+++ src/components/ComposedModal/index.js
@@ -1,0 +1,1 @@
+export default from './ComposedModal';
