+++ src/components/WizardModal/WizardModal.jsx
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import isNil from 'lodash/isNil';
 
-import BaseModal from '../BaseModal/BaseModal';
+import ComposedModal from '../ComposedModal/ComposedModal';
 import ProgressIndicator from '../ProgressIndicator/ProgressIndicator';
 import WizardFooter from '../WizardInline/WizardFooter/WizardFooter';
 
-const StyledModal = styled(BaseModal)`
+const StyledModal = styled(ComposedModal)`
    {
     .bx--modal-container {
       min-width: 410px;
@@ -40,7 +40,7 @@ const StyledWizardContent = styled.div`
 `;
 
 /**
- * Extends BaseModal to add wizard, refer to that component for common props
+ * Extends ComposedModal to add wizard, refer to that component for common props
  */
 class WizardModal extends Component {
   static propTypes = {
