+++ src/index.js
@@ -5,7 +5,7 @@ export styles from './index.scss';
 export ButtonEnhanced from './components/ButtonEnhanced';
 export Table from './components/Table';
 export AddCard from './components/AddCard';
-export BaseModal from './components/BaseModal';
+export ComposedModal from './components/ComposedModal';
 export WizardModal from './components/WizardModal';
 export WizardInline from './components/WizardInline/WizardInline';
 export StatefulWizardInline from './components/WizardInline/StatefulWizardInline';
@@ -47,7 +47,6 @@ export {
   Checkbox,
   CodeSnippet,
   ComboBox,
-  ComposedModal,
   ModalHeader,
   ModalBody,
   ModalFooter,
